requirements.txt
langchain-experimental>=0.3.4
langchain>=0.3.19
openai>=1.63.2
pandas>=2.2.3
plotly>=6.0.0
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
firebase-admin>=6.0.0
trafilatura>=2.0.0
langchain-community>=0.3.18
twilio>=9.4.6
tabulate>=0.9.0
openpyxl>=3.1.5
pandasai
fastapi
uvicorn
mangum
python-multipart
langchain-community
